refactor(tui): extract event dispatch and effect processing into handlers
Move all event dispatch logic and background action processing from
run() into handlers.rs. This includes modal handlers (session manager,
model picker), mouse events, keyboard dispatch for Input/Cursor modes,
and the cancel-generation helper.

The active_abort_handles Vec moves into TuiState since it's TUI state
that the cancel helper needs mutable access to.

After extraction, run() is a thin loop: setup → draw → poll →
handle_event() → process_background_actions() → teardown.

mod.rs: 793 → 224 lines.
